storage: make tracking of storage dependencies more principaled
Before, we were tracking for each collection any sinks/exports that
depended on that collection. And we had somewhat complicated code that
was making sure we take their requirements into account when calculating
a new since for a collection.

Now, we turn that relationship around. Exports register requirements on
the since of their dependencies (the collection that they are sinking
from). And we track those dependencies and forward any changes in
requirements. Similar how the compute controller tracks since changes
between collections (including storage collections).