sql: Move statement size check to server layer (#17580)
Previously, the validation of SQL statement size was done in the
parser. The issue is that the parser is used by bootstrap to read in
catalog items. So if a catalog item was somehow committed that was
larger than our max size, then the database will fail to start up. This
can legitimately happen for various reasons. For example, a customer
created a view before this check was added or the planner increases the
size of a statement after the check happened.

This commit moves the statement size check to the pgwire, http, and
web-socket server layers. This ensures that bootstrap will never fail
with statements that are too large, and the check is only done on
statements sent directly from a client.

Fixes #17573