persist: add a referenced blob closure to state iter's next
A number of places currently compute some set of referenced blobs
(batches and rollups) using a hash or btree set. This is `O(seqnos *
blobs)` work, but that's a bit silly/wasteful because we already store
things as diffs. This commit introduces a closure that gets called once
per blob as we iterate through versions of state, so the above can be
done in `O(blobs)` work and without the set.