persist: introduce additional granularity into usage tracking
This is in service of changing which class of things we bill for. This
commit introduced new breakdowns into the usage calculation... and then
throws it away and returns the old numbers. Changes to the caller of
usage code will be done in a followup PR, since that wants an entirely
separate set of reviewers. See #17317 for context.

Closes #17293