Eliminate panic from uint2 and uint4 sums under invalid accumulations
This commit eliminates a panic raised when a sum aggregation of uint2 or uint4 is faced with a negative accumulation. Negative accumulations in reductions can arise due to, e.g., incomplete or invalid source data. The new strategy is to convert the invalid aggregation error into a query-level error instead of a panic by introduction of an MFP that performs appropriate casts and projections. The MFP is currently introduced at the level of mir_to_lir lowering, since the concern is closely connected to how rendering deals with invalid accumulations. The latter avoids needing to change how aggregate function evaluation is performed for constant folding, but comes with the trade-off that the MFP cannot be pulled up by the optimizer to be evaluated at a later stage in the plan.