compute*: avoid FrontierUppers regressions
Prior to this commit, it was possible for replicas hydrating a dataflow
to report new upper frontiers that were below the dataflow `as_of`. To
the compute controller this would look like a frontier regression, and
it would potentially become confused.

This commit prevents that confusion be implementing three measures:

 1. Making replicas suppress upper updates for frontiers that are less
    than the respective dataflow `as_of`.
 2. Changing the compute protocol spec to disallow regressions in
    `FrontierUppers` responses.
 3. Adding a check to the compute controller to detect and gracefully
    handle frontier regressions (by ignoring them).

Measure (3) is not strictly required and only acts as a safeguard
against unintentional violations of the compute protocol.