Sentry error logging and regression testing for invalid accumulations in unsigned sum aggregations
This commit adds Sentry error logging to reduce rendering so that we can become aware of when an unsigned sum aggregation results in a negative number. The cause for this odd behavior is invalid source data with too many retractions. In addition to error logging, this commit introduces a regression test that ensures that: (a) We produce SQL-level errors when we hit this situation as well as log error messages; (b) The types produced at the output of unsigned sum aggregations are the expected ones.