persist: introduce idle_merge_effort analog
In differential dataflow, idle_merge_effort configures an arrangement to
introduce extra fuel on every operator invocation. This causes data to
continue compacting, even when new updates are not being introduced.

Persist doesn't have anything that directly corresponds to operator
invocations, but we do the best we can to match the spirit of
idle_merge_effort. Specifically, we introduce the additional fuel on
each compare_and_append call. In practice, compare_and_append gets
called on a regular cadence to advance the frontier, so even in no
updates are being added, we'll eventually compact things down. This does
mean, however, that tuning the constant involved will likely be quite
different.

Touches #15093
Touches #16607