compute*: avoid subscribe frontier regressions
Prior to this change, it was possible for replicas hydrating a subscribe
to report batches with `upper` frontiers that were below the subscribe's
`as_of`. To the compute controller this would look like a frontier
regression, and it would potentially become confused.

This commit prevents that confusion by implementing three measures:

 1. Making replicas emit only subscribe batches that have `upper`s
    beyond the respective subscribe `as_of`.
 2. Changing the compute protocol spec to disallow regressions in
    subscribe batch `upper`s.
 3. Adding a check to the compute controller to detect and panic on
    write frontier regressions.

Measure (3) is not strictly required and only acts as a safeguard
against unintentional violations of the compute protocol. Note that in
contrast to `FrontierUppers` we cannot simply ignore subscribe batches
that report frontier regressions, as that would change the subscribe's
output.