repr: add warning about GlobalId's Ord implementation
It was previously true and thought to be the case that GlobalId's
Ord implementation expressed some invariants about entries'
dependency graphs. However, in a future commit, we will definitlvely
break this invariant and it cannot/should not be relied on.

However, we cannot simply remove the Ord implementation, which would
prohibit building sets/maps from GlobalId, which is far too common
of a pattern for us to prohibit.