sql: apply AST transforms to all expressions
The SQL planner was failing to call `transform_expr` on several
expressions: `DEFAULT` expressions in `CREATE TABLE` statements and
`RETURNING` expressions in `INSERT` statements. This could lead to
panics when e.g. the expression contained an `Expr::Nested`, which the
planner expects to be transformed away by `transform_expr`.

This class of bug has bitten us several times. We should attempt a more
robust solution (e.g., using the Rust type system as @ggevay suggests in
[0] to distinguish between untransformed and transformed expressions),
but doing so requires a substantial refactor that I don't have time for
at the moment.

Fix #17706.

[0]: https://github.com/MaterializeInc/materialize/issues/17706#issuecomment-1434286617