stash: improve error retry heuristics
The linked issue observed a stash with a diff of -1 for a catalog item
while scaling down a cockroachdb cluster. This could have happened after
cockroach committed a transaction but before the response was able to
get back to the stash. The stash incorrectly attempted to retry,
resulting in applying the delete twice.

Improve the stash retry heuristics by annotating which part of the
transaction failed, since they all have different retryability
semantics. To prevent this specific error, the COMMIT retry is greatly
limited to only being retryable when cockroach specifically requests a
retry.

Fixes #17754