Introduce validating reductions at the MIR level
This commit introduces an attribute in MIR Reduce nodes that indicates if the given reduction produces an extra validity column. Such a validating reduction would check its input for each key to detect invalid accumulations and then produce a corresponding boolean result in this extra validity column. The commit is presently inert in the sense that all reductions are marked as non-validating. However, it serves as a mapping of the extent of the changes that would be necessary to our optimizer transforms to support the concept properly. Note that the commit does not include hooking up validating reductions all the way to rendering.