sql: Create roles if they are authenticated (#17873)
Previously, if a user successfully authenticated through Frontegg, and
they didn't have a role in Materialize, then we would create a new role
for them. This commit updates this logic so that if a user successfully
authenticates through whatever authentication method the Materialize
deployment is using, then we create a new role for them.

This allows us to remove the `materialize` user, without breaking local
development. Once RBAC lands in #11579, the `materialize` user presents
a potentially scary security vulnerability since it defaults to having
all privileges. It's also only used for local development.

The HTTP and WS API already worked like this, so only the pgwire
protocol needed to be updated.