storage: send `AllowCompaction` commands for sinks/exports
This will make sure that potential changes to the `as_of` of a
sink/export (which is used when restarting) are reflected in the command
stream. Which will then allow us to get rid of brittle code that asks
persist about sinces in both the rehydration code and storage worker
code.