adapter: Derive superuser attribute from auth
This commit updates the startup command to derive the superuser
attribute of a role. Each time a connection starts, the superuser
attribute is derived by checking if the user is an admin in the
external authentication system or is the `mz_system` user.

This commit will also update the superuser attribute of a connected
session every time their authentication is refreshed by an external
authentication system.

Part of #11579