timely-util: more ergonomic async handle
The pattern of managing an operator local buffer that gets swapped with
every `RefOrMut` received is pervasive and only adds noise.

This PRs adds this feature to the handle itself so that users can call
either `.next()` and get a normal reference or `.next_mut()` and get
mutable reference to a pre-swapped container, eliminating the
boilerplace.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>