Only validate invalid accumulations in the first reduction of a bucketed aggregate
This commit changes the validation policy of invalid accumulations in bucketed aggregates. Previously, invalid accumulation checks were performed at every level of the hierarchy. However, this was redundant: A failed check at the first level would lead to the same detection at all further levels of the hierarchy. With the change, we perform the check at the first level of the hierarchy only, saving resources and reducing noise in the error logs.