Partial regression test for issue #17509
This commit adds a partial regression test for the new behavior of bucketed aggregates subjected to invalid accumulations. The test is part of the mzcompose cluster workflows so that we can check that the appropriate error messages are produced in the logs. We issue queries in which a hierarchy of reductions is built, in which case invalid accumulations can produce a clean query-level error. We also test that resolving the error is possible by correcting the base data.