storage: order independent bytestream decoding
We were previously relying on timely channels maintaining the order of
messages but this is not a guarantee offered by timely and has led to
bugs in the past.

This patch introduces a `ByteStream` type that all sources like S3 must
use which encapsulates the fact that we want to treat each byte stream
as a single "datum" but we don't really want to buffer it in memory at
once.

The downstream non-delimited decoder receives `ByteStream` instances and
can consume them in order without relying on the chunks being received
and reassembled over timely channels.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>