storage: make schema change detection more robust
We were previously only determined schema equality using column
names, which made it impossible for us to detect if a schema's
columns were changed but retained the same ordering of names.

To resolve this issue while in the replication stream, we must
contact the upstream source to determine schema we've been notified
about. Unless PG adapts the Relation replication stream message
to unamiguously identify the schema columns, this is inevitable,
unless we want to blithely error on all schema changes.