ci: Clean up all docker containers on job start, regardless of origin
Previous pre-job cleanup was either kind-only or
mzcompose composition-specific. This caused CI hosts
to accumulate old running mzcompose containers which
in turn could lead to OOMs and all sorts of other
issues for the CI job being run.

In order to collect more information about the situation
in the CI, the new cleanup is explicitly set to run in
addition to any existing cleanup actions even if it is
more general than them.