slt: script for generating cluster_log_sinks.slt
This commit moves the Python code that generates cluster_log_sinks.slt
from a comment in that file into its own Python script. This reduces the
hassle induced by having to make changes to and then running the
codegen.

This commit also uses the opportunity to clean up the codegen code a
little.