adapter,compute: unify introspection histograms
This commit unifies the existing compute introspection histogram sources
by:
 * Adding the suffix `histogram` to their names.
 * Removing the `sum` field from all of them.

The latter change means that we now can expose `_internal` sources that
have the `count` field exploded for each histogram. This is makes
logging for these sources a bit simpler and possibly also more
efficient.

The `sum` fields were initially introduced to enable exporting
histograms to Prometheus. However, it is not clear how much value
computing this extra field has over approximating it based on the
bucket value and `count` field. So we decided to remove it for now.