adapter: Add role membership
This commit adds the ability to add one role to another role.
Additionally, it adds the ability to remove one role from another. If
role_a is a member of role_b then role_a inherits all the attributes
and privileges of role_b.

When a role is dropped all other roles are automatically revoked as
members from that role.

Part of #11579