persist: plumb schemas to all necessary places
Notably, we need different schemas for encoding vs stats in compaction
(due to the possibly questionable choice of compaction working on the
raw bytes). Fixing compaction to operator on the real K and V would have
the potential for all sorts of subtle consequences, and if we commit to
this strategy, the old codepath is going away. So, instead we split a
BatchBuilderInternal out of BatchBuilder, which specifies everything
except the schemas to use for stats. The public BatchBuilder just plugs
in the same schemas for encoding and stats, but compaction can drive
around BatchBuilderInternal directly.