storage: work with `Collection` everywhere
Many of the transformations in the ingestion pipeline should be
expressed as transformations on collections without peeking into the
underlying timely stream. This forces us to think about our
transformations purely on the data of the collection, acting as a
guardrail to stay in the IVM mode.

This PR is mostly mechanical and only changes signatures to work with
`Collection` and only converts transformation steps when they were truly
trivial.

In a subsequent PR we should attempt to translate more of our operators
(decode looks like an easy start) to use `Collection::map` or
`Collection::flat_map`.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>