storage: remap: keep up with progress updates
The remap operator has a starvation issue where if `select!` is totally
fair and minting a binding takes more time than the ticker interval then
the next time we go around the loop we might peel off one or two
progress updates before entering the ticker again and minting a binding
for a tiny slice of the available progress. In this situation we keep
falling behind in minting bindings causing data to be accumulated in the
reclock opreator which sits on data that need to be reclocked.

This PR makes it so if we're falling behind we will make bigger and
bigger jumps in the bindings by consuming the entirety of the available
progress when there is any.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>