compute: simplify worker handling in compute logging
Logging events are always emitted on the same worker that handles them.
So there is no need for the demux operator to shuttle the worker ID
through together with each event -- it is always the same as the current
worker's ID. Instead, we need to insert the worker ID only when we
encode the final output `Row`s.