persist: initial skeleton of part stats collection and filtering
This is controlled by two independent DynamicConfigs. When set,
`stats_collection_enabled` will compute and durably record PartStats for
each user and compaction part. When set, `stats_filter_enabled` will use
any PartStats that are available to skip fetching parts at read time.
PartStats are never required, so either config may be set without the
other.

This is intentionally a minimal skeleton of collectings stats on parts
and using them to skip fetching parts entirely in persist_source. This
will allow productionized versions of the various pieces to be worked on
with far less coordination. It includes a number of limitations:

- This commit/PR only introduces changes in src/persist*, changes to
  persist users for the skeleton will be done in followup PRs to enable
  tighter code review audiences. As a result:
  - The configs are not yet hooked up to LaunchDarkly.
  - The impl of `Schema<SourceData> for RelationDesc` remains a TODO for
    now.
  - The filter is not hooked up to MFPs.
- Many `TODO(mfp)` comments are left. These are things that certainly
  will be fixed before turning this on in prod and have the additional
  suffix to allow for easy grepping.
- Stats are only collected for `Option<u64>` columns.
- Stats are stored in metadata in an extremely inefficient way (one per
  ProtoHollowBatchPart).
- Stats are computed for parts in the old format by wastefully
  re-encoding them to the new (schema'd) part format, computing stats on
  that, and then throwing it away.
- No metrics are kept.
- No tests are included.

A slight performance hit is expected due to additional Arc::clones from
plumbing the Schemas around. We'll let Feature Benchmarks decide if this
is significant.