persist: placeholder impl of `Schema<SourceData>` for RelationDesc
This implements Schema for SourceData by flatmap-ing the Ok Row's
columns and the Err. This has the unfortunate effect of requiring us to
make all Row columns nullable (even if they aren't in the RelationDesc)
so we have something to store if the Err column is set. (Luckily, we can
still check it at decode time.)

Better would be something like pushing the union structure of Result
down to parquet, but that's much harder and left for followup work (if
we do it at all).