storage: refactor UPSERT ingestion pipeline
We modularize things, pull the code that extracts key, value, and other
things from a DecodeResult out into its own operator. Plus we add some
comments.

This tries to make it so that we only use `UpsertError` throughout the
UPSERT operators while we get as input `DecodeResult` (which can contain
`DecodeError`, and have to have `DataflowError` in our "current values"
map and as output. See also comments in code.

Will will make it easier to add things like handling of NULL keys
easier, which we are planning to do in a next commit.