storage: stop silently ignoring NULL keys in upsert source
Before, we were logging an error and continuing, which, to the user, is
"silent".

Now, we report an error when trying to query the source. And we support
retracting them _too_.

There is some nuance to what we can feasibly report as the error. I have
this `NOTE` in the code:

> NOTE: We need to make sure that when we see a "retraction" for this, that is, a
> message with _both_ a NULL key and NULL value, we can match this to the original
> `DecodeError`. Therefore, we cannot put more information in here than the
> partition. It would seem desirable to have more information, for example the
> offiset, but with that the future "retraction" message will not match this
> message and the original error will not go away.

Fixes #6350

As follow-up work, we might want to do more involved things with NULL
keys.