testdrive: Reset Materialize with system user
Previous commits have added partial implementations of Role Based
Access Control (RBAC), and future commits will continue to add more
RBAC features. As part of RBAC, ordinary roles may be prevented from
executing certain actions, such as creating/destroying databases.

This commit modifies testdrive so that resetting the state of
Materialize is always done with the system user. This is to ensure that
we are never prevented from resetting Materialize's state due to RBAC.
The system user is always a superuser and is therefore always allowed
to do anything.

Part of #11579