compute,adapter: remove dedicated sink `as_of`
This commit removes the dedicated `as_of` frontier that is part of
compute sink descriptions (it does not touch storage sinks). Our code
sometimes makes the assumption that the sink `as_of` is greater than or
equal to the dataflow `as_of` but does not thoroughly check that
assumption. The approach taken here is to remove the sink `as_of` and
rely on the dataflow `as_of` in its stead, thereby also removing the
need to worry about the stated assumption.

The advantage of this approach is that we don't need to introduce
logic for checking and maintaining the above invariance.

The (AFAICT only) drawback of this approach is that it might make it
harder to support dataflows that export multiple sinks in the future. If
we want to do that at some point, and if we then need the feature that
different sinks can have different `as_of`s, we will need to
re-introduce the per-sink `as_of` again.