adapter: Put inner sequence method in private mod (#18149)
This commit moves all inner `sequence_X` into a private module. Only
the `sequence_plan` method will be visible outside the `sequencer`
module. This helps guarantee that all callers must call `sequence_plan`
and can't call an inner `sequence_X` directly. This allows us to put
catch-all logic in `sequence_plan` and guarantee that it will be
applied to all plans.

The one exception is creating a role during connection startup. In this
scenario the user's session has not been initialized and the checks in
`sequence_plan` are not applicable.