sql: correctly handle function selection with unknown types
Our function selection code had a buglet in which the metadata required
to assess case 4f was not being correctly computed, because the
computation of that metadata was intertwined with the assessment of case
4e.

This commit fixes the bug by splitting 4e and 4f into completely
separate loops. Case 4e is now free to break early, without impairing
the computation of case 4f's metadtaa.

The user-facing impact of this change is that a SQL query like

    SELECT '1' + '1'::int

now plans correctly. Previously it was rejected with an "operator is not
unique" error.

Fixes #18240.