storage: improve SSH tunnel failure errors
We've been seing errors like

    failed to close ssh tunnel: the master connection failed

in production. The actual *reason* that the SSH tunnel failed is buried
inside the returned error type, and not part of the `fmt::Display`
representation.

This commit adjust the logging to route through `anyhow::Error`, which
can be convinced to print all underlying error causes via the `{:#}`
format specifier.