ci: Standardize Nightly Zippy execution times to 55m
Previously, each variant of Zippy was running for a different
amount of time in Nightly.

Use 55m rather than 60m in order to avoid machines being billed
for a second hour if it is not going to be used.

The exception is the "Cluster Replicas" step, which can not
currently run for 55 min.