slt: Fix default cluster clean up (#18306)
The syntax of deleting cluster replicas requires specifying the cluster
name and the replica name. For example, `DELETE CLUSTER REPLICA
default.r`. However, when SQL Logic Tests was deleting cluster replicas
in the default cluster, it was failing to specify the cluster name,
causing the cleanup process to fail.

This commit fixes the described issue.