sql: check signed division for overflow
The case INT_MIN / -1 is equivalent to (-INT_MIN) which is one larger
than INT_MAX.

INT_MIN % -1 overflows even though it equals 0, because at least on
x86, remainder is computed with `idiv`, which raises `#DE` on
overflow (which turns into SIGFPE on Linux, explaining why this also
happens in release builds).

We return 0 anyway (the mathematically correct answer) because it's
what Postgres does:

  https://github.com/postgres/postgres/blob/adedf54e659ea37316a21c8d5a7aa8f711213d1c/src/backend/utils/adt/int.c#L1144-L1150

Note that in these cases, since we have already checked the modulus
isn't 0, overflow is the only reason for checked_rem to fail.

Fixes #18181.