sql: Refactor ObjectType enum
The ObjectType enum, which is used for parsing and planning, seemed to
be confused as to what the definition of object was. The two reasonable
definitions would be:

  1. Anything the exists within a schema, what Materialize refers to as
  a CatalogItem or CatalogEntry.

  2. Any logical database object that a user can create or drop.

The fact that ObjectType included roles, clusters, and cluster replicas
excluded it from category (1). The fact the ObjectType didn't include
databases and schemas excluded it from category (2). The current
definition seemed to be any logical database object that a user can
create or drop and isn't a namespace.

This commit adds databases and schemas to ObjectType in order to
conform to definition (2). I chose definition (2) over (1) for the
following reasons:

  - PostgreSQL already refers to these as objects and identifies these
  with oids (Object IDs).
  -  Currently all code that uses ObjectTypes also considers databases
  and schemas.

I was able to simplify most code as a result of this refactor. Some
spots needed a special cases for databases and schemas, but there
were already special cases for roles, clusters, and cluster replicas.

This commit also removes the Object variant from ObjectType because I
found it to be a bit of a wart. It was only used in a single place and
complicated all other uses of ObjectType.