storage: refactor shard finalization
We previously asked that we be presented with the set of shards
to finalize; however, we can instead idempotently process the
WAL we retain that contains the list of all shards we want to
finalize.

We can also defer the work until we see that there are likely to
be shards to finalize, which can decrease the time we take on boot.