adapter: Add object owners (#18280)
This commit adds logical owners to the following types of database
objects:
  - sinks
  - indexes
  - connections
  - types
  - functions
  - secrets
  - tables
  - sources
  - views
  - materialized views
  - databases
  - schemas
  - clusters
  - cluster replicas
  - schemas

All objects that exist at the time of a new Materialize deployment
(including all system objects) are owned by the mz_system role. Any
object that is created by a user is owned by the creating user.

This commit also adds a migration that adds a "default_owner" role and
assigns that role as the owner of all existing objects.

Currently, owners are not used for anything. In a future commit we will
add privileges based on an object's owner, and we will add the ability
to update an object's owner.

Part of #11579