Make replica IDs strings in the system tables
This commit changes the type of replica IDs in system tables (i.e.
`mz_cluster_replicas.id` and foreign keys referencing that column) to a
string in `GlobalId`-compatible `NewReplicaId` format.

The intention of this is to prepare for a later introduction of system
replicas, which can be marked as such by having a system ID.

The change is intentionally kept as small as possible to be easily
revertable in case we discover during the release that customers have
created last-minute dependencies on `mz_cluster_replicas`. For this
reason, actually using `NewReplicaId` as the type for replica IDs
throughout the code is left as future work.