stash: Revert serialized role struct (#18418)
In a previous commit `RoleAttribute` was added to the `SerializedRole`
struct. It was originally added as an `Option` and accompanied by a
migration to always set the value to `Some`. After 3 releases, we
removed the `Option`. from `RoleAttribute`. Due to a bug in the stash
that failed to consolidate old entries, old values of `SerializedRole`
with a null `RoleAttribute` field existed in the stash. The lack of an
option on `RoleAttribute` caused deserialization of these null fields
to fail, which prevented environmentd from starting up.

This commit adds back the `Option`, so environmentd can start up
successfully. The underlying consolidation bug will be fixed in a
future commit.