storage: define `SourceRender` trait
This commit defines the `SourceRender` trait which supercedes all the
previous ways a source could be defined. Sources are now able to render
arbitrary dataflows, taking advantange of multiple communicating timely
operators when needed.

A lot of the auxiliary types that used to support the old traits have
been removed.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>