storage: add option for pre-reducing (or not) UPSERT updates
Pre-reducing matches what we have in production, but this also largely
benchmarks the additional BTreeMap that we use for pre-reduction.

Also, if the key cardinality matches the number of records per second,
this degenerates to doing double the number of BTreeMap inserts.

The non-pre-reducing version is harder on the state backend, and should
give us a clearer signal.