Add MaybeValidatingRow
Between #17918 and the changes for #17963, it's clear that a pattern
is emerging around validating in the first stage of a multistage
operation.

This trait is an attempt at abstracting the output type of reductions
so that validation can be added relatively unobtrusively to such
stage-building code.

Kudos to Jan for the idea to make the error constructor optional.

Co-authored-by: Jan Teske <jan@materialize.com>