persist: calculate and store stats for all `T: Data` types
This expands the set of column types that persist internally whitelists
stats to all types (it was previously only Option<u64> in the skeleton).
This might considerably increase the size of the stats metadata, but
it's behind a feature flag.

Combined with #18386, we'll get useful stats for a much larger number of
ScalarTypes. Finishing off the last few will require a Schema impl to
override the stats logic (e.g. for schemaless json stats, understanding
Decimals, Lists, etc) is left for followup work. In the meantime, we'll
store meaningless stats about their ProtoDatum representations. Actually
using these new stats in MFP pushdown is also left for a followup PR.

To accomplish this, we introduce column stats as a first-class concept
in mz_persist_types. It's possible this ends up being unnecessary, but
it seems like a likely enough bet and we can always pull them back into
mz_persist_client if it turns out to be wrong.