testdrive: fix subscribe frontier test
This test runs a `SUBSCRIBE` that is supposed to verify that its own
frontier is advancing in the introspection sources. Previously, it
simply looked for its own frontier by searching for dataflow exports
with a "t%" ID. Unfortunately, transient GlobalIds are not only given to
subscribe dataflows but also to ad-hoc `SELECT` dataflows. As a
consequence, this test was likely to check the logged frontier of one of
the previous `SELECT`s, rather than that of the subscribe. This is now
fixed by looking the subscribe ID up in `mz_subscriptions`.