compute: avoid logging the logging errs arrangement
Prior to this change, the empty arrangement providing the `errs` part of
the trace bundles for logging collections was always created after the
timely loggers had been registered. As a consequence, there would always
be one entry in `mz_arrangement_sharing` on an empty replica, even if
the replica was created without "log_logging" enabled.

This commit fixes this by moving the creation of the logging errs
arrangement into the same function that creates the logging dataflows.
Thus, this arrangement is now only logged when creation of the logging
dataflows is also logged.