Handle errors in ReduceCollation
This is an uncomfortable change because the error logic needs to live
in a separate closure, and unlike some of our previous uses of this
pattern, the main logic closure is large enough that it's hard to
follow both at once.  Still, after much deliberation, this seems like
the only reasonable way to surface these errors as requested in
the associated issue, #18215.

Fixes #18215, #18216.