sql: Add maclitem type for privileges (#18513)
This commit adds the maclitem SQL type which is used to store role
privileges. This type is modeled after the aclitem type in PostgreSQL.
We were not able to reuse the aclitem exactly, because aclitem uses
OIDs to reference roles, while Materialize uses a custom type called
RoleId to reference roles. The 'm' in 'maclitem' stands for
Materialize.

The PostgreSQL aclitem implementation can be found here:
https://github.com/postgres/postgres/blob/7f5b19817eaf38e70ad1153db4e644ee9456853e/src/include/utils/acl.h#L48-L59

Part of #11579