storage: let sources provision external dependencies early
We want to permit users to name their own PG replication
slots; to support this, we want to ensure that we are only
letting users name new PG replication slots, and do so
without TOCTOU issues.

The easiest way to do this is check as late as possible
in the process of creating a source that we are creating
a new replication slot; easiest in this way because we
also want to drop the replication slot if we fail
for any reason.