*: replace alternate selector by `display_with_causes()` when printing errors
Before, we were relying on the `Display` impl for our errors to actually
print the error chain. This is true for `anyhow` errors but not for most
other types of errors.

Now, we explicitly request that we want to print the error along with
its causes.

This covers all places I could easily find, there could be more that we
have to find over time.