ore: change ResultExt to use error-with-chain printing
This is the last vestige of code that was relying on the _alternate
selector_ for printing `anyhow::Error` with it's chain of source errors.
We previously introduced `ErrorExt`, with methods that allow explicitly
requesting the full error chain be printed, for _all_ types of errors
(`std::error::Error`) that implement `source()` correctly.

This also renames the methods to make the `_with_causes` nature more
explicit.

And we remove `DisplayExt`, which was by now an implementation detail of
the previous incarnation of `ResultExt` and which was a potential
footgun because it would, again, only print the chain of errors for
`anyhow::Error`.