adapter: Buffer updates to metrics system tables
Previously whenever we needed to update a system table containing
metrics, we would immediately trigger a potentially expensive group
commit whenever a metric was updated. Metrics are not critical enough
that we needed to update the system table immediately.

This commit update the append logic so the table updates for metrics
are buffered but don't trigger a group commit. Instead, they will
piggyback off the next group commit. This will hopefully allow multiple
metrics updates to be coalesced into a single group commit.

This commit also restarts the table advancement timer whenever a system
table is updated, which was left off of a previous commit. In the
previous commit only user writes would reset the timer.