pg sources: skip deduplication from snapshot if possible
If there are no changes in the WAL and we enter produce_replication,
we wait for 30 seconds until we produce a progress event.

This is annoying in testing environments when we're more likely
to have to enter produce_replication for snapshot deduplication and
are less likely to have content in the WAL that can produce
Progress messages at times beyond the snapshot. Receiving a progress
message at an LSN beyond the snapshot is required to serve any reads.

To resolve this, we can peek into the WAL for any LSNs between
our current position and the snapshot LSN; in the case where there
are none, we can avoid entering produce_replication for deduplication
because there is nothing to deduplicate.