repr: don't use ResultExt in strconv
It was using `ResultExt::map_err_to_string` which was introduced for
printing errors along with their chain of `source()` errors. However,
none of the involved errors do actually have source errors, and we were
doing the printing of the chained errors (here called _details_) by hand
already, in the `Display` impl of `ParseError`.

I believe this was used purely out of the convenience of typing
`map_err_to_string()`. We will change that however, so should make what
we do in strconv more explicit.