adapter: Buffer updates to metrics system tables (#18622)
Previously whenever we needed to update a system table containing
metrics, we would immediately trigger a potentially expensive group
commit whenever a metric was updated. Metrics are not critical enough
that we needed to update the system table immediately.

This commit updates the append logic so the table updates for metrics
are buffered but don't trigger a group commit. Instead, they will
piggyback off the next group commit. This will hopefully allow multiple
metrics updates to be coalesced into a single group commit.