Introduce Row::kind_iter
Introduces an iterator over the datum kinds within a Row, which can be used
instead of full datum decoding where merely the type or existence of a
datum is important.

Decoding the datum's kind requires reading tag values and length data for
variable-sized kinds, but can skip over actual data.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>