adapter: Add privileges to objects
This commit adds privileges to the following objects:

  - tables
  - views
  - materialized views
  - sources
  - types
  - secrets
  - connections
  - secrets
  - clusters
  - databases
  - schemas

This commit updates the on-disk stash representation and the catalog
tables/views that present this information to users.

Currently, privileges cannot be modified, and they are not looked at
when executing statements. These features will be implemented in a
future commit.

Part of #11579