adapter: Refactor cascading delete statements (#18651)
Previously, the adapter would collect all the IDs of objects to delete
as part of a cascading delete during sequencing. This commit refactors
this logic so the IDs are collected during planning. This allows us to
include those IDs as part of RBAC checks.

Additionally, we had to move the validation that no user was directly
modifying a linked cluster to the planner.

Fixes #18551