pg sources: fix no-progress replication bug
When adding a timeout for the replication strea, we inadvertently
broke from the replication loop when exceeding the deadline. This
meant we would re-enter the loop starting at the same LSN, making
it possible that we would never make any progress.

Instead, when exceeding the deadline, skip the message handling
and go straight to the status update.