environmentd: allow TLS connections to the internal SQL port
When the external SQL port requires TLS, allow TLS connections to the
internal port as well. This allows tools like Teleport, which insist
upon using TLS for all connections, to connect to the internal SQL port,
without breaking existing users of the internal SQL port, which may need
to be reconfigured to support TLS encryption.