pg sources: do not fast-forward on KeepAlive if pending writes
The previous code was written under the assumption that we waited
indefinitely on the replication; however, now that we periodically
break from waiting on the stream, we need to be more mindful of
managing the buffered inserts/deletes.

This ended up manifesting as a bug where we had pending writes,
received a keepalive message, and then reconnected at the same
original LSN, meaning we ingested the same data multiple times.

Instead, do not try to fast-forward if there are pending writes.