compute: don't log errors during dataflow shutdown
This commit modifies the error logging behavior of compute operators to
only log messages when the dataflow is not in the process of shutting
down. This avoids producing noise caused by dataflow operators emitting
partial update sets when they are cancelled.

The shutdown checking is performed through a shutdown token and wrapped
inside the new `ErrorLogger` type that gets passed around to the various
operator renderers.