sql: Ensure role member grantor is always valid
This commit prevents dropping a role, if that role is a grantor fo some
other role membership. This helps ensure that the grantor references in
the catalog always remain valid.

As of this commit, the only valid grantor is mz_system which is
impossible to delete. So it's impossible to hit this error scenario.
However, this is more future-proof for when we implement ADMIN OPTION
and allow other grantors.

This is based off of the equivalent commit in PostgreSQL here:
https://github.com/postgres/postgres/commit/6566133c5f52771198aca07ed18f84519fac1be7
That commit will be part of PostgreSQL v16 and was not included in v15.

Part of #11579