sql: Update role membership for PostgreSQL v16 (#18780)
Previously, role membership was modeled after the implementation of role
membership in PostgreSQL v15. Specifically the grantor was set as the
role ID of the session executing the `GRANT` statement. However,
PostgreSQL v16, which has not been released yet as of this commit,
updated role membership to be more consistent with the SQL standard and
with object privileges. Specifically the grantor could only be set to
someone with ADMIN OPTION on the role being granted or the bootstrap
superuser.

This commit set's the grantor of all role membership to the mz_system
role. We haven't implemented ADMIN OPTION yet, and the 'mz_system' role
is our version of the bootstrap superuser.

The PostgreSQL commit that made this change can be found here:
https://github.com/postgres/postgres/commit/ce6b672e4455820a0348214be0da1a024c3f619f

Part of #11579