doc/user: Refresh the Getting Started guide
This PR revises the "Get started" guide to walk users
through the main steps in the journey, introducing
important Materialize concepts along the way. It also
uses a load generator as source instead of a managed
Kafka cluster.

This PR also adds a page with FAQs for the free trial
and adds `psql` to the integrations overview.

Fixes: #18417
Fixes: #18954
Fixes: #18882