timely-util: don't shut down progress in `with_token`
The previous implementation of `with_token` progress handling was flawed
and prevented downstream operators from seeing any progress. In
addition, there is some uncertainty about the safety of aggressively
advancing to the empty frontier on token drop within dataflow loops. So
this commit reverts back to the previous behavior of only dropping data
updates on token drop. We can add the frontier advancement back in as a
follow up, if we agree that it is safe.