Add test for `oom-killed` reason in `mz_cluster_replica_statuses` (#18806)
This test creates an index on a query that is known to OOM, repeatedly
checks `mz_cluster_replica_statuses` until the OOM is observed, clears
the query, and then repeatedly checks until the cluster is observed
ready.

It's possible to imagine that this test can have false negatives (i.e.,
the test is green but the underlying behavior is broken) if certain
patterns of flapping occur in the data. I think this is unlikely to be a
concern in practice. It would be more precise/elegant if we could
control how long it takes a pod to restart after OOMing, but there
doesn't currently seem to be a way to do so.