test: Ignore drain timeouts during a CRDB rolling restart test
Sometimes crdb's drain command is unable to drain all the connections
from a node. Ignore this when it happens and proceed with the rolling
restart as it will be assumed that in practice, during a rolling upgrade
in production, the same sequence of events will happen -- a drain timeout
will be ignored and the upgrade will proceed.