doc/user: add documentation for inline Protobuf schemas
We've long supported specifying inline Protobuf schema descriptors,
which is required for using Materialize with Protobuf-formatted Kafka
topics that do *not* use the Confluent Schema Registry.

Unlike the Avro format, the Protobuf format is self describing, so using
Protobuf without the CSR possible to do correctly, if a bit annoying.