sql: don't advance AS OF AT LEAST to upper
Previously we would advance AT LEAST to the upper because...we could? It
improved freshness while still meeting the requirement of AT LEAST,
which is that the specified timestamp is included in the output.
However, this interacts poorly with SUBSCRIBE UP TO, because the upper
might be in advance of the UP TO, causing an error.

Instead use the oldest valid timestamp for AT LEAST. The meaning has
changed to: get the data as close as possible to the specified
timestamp. This allows for historical queries to return more detailed
data.

Fixes #18910