adapter: re-verify and plan off-thread statements
Statements that can execute off of the coordinator thread have the
potential to become invalidated if their underlying objects or indexes
have changed while doing work off of the coordinator thread. Detect this
by recording the transient revision and detecting if it changed. Add a
common way for statements to re-plan themselves from the beginning when
this happens.