pg sources: propagate definite source errors more carefully
In upgrading to PG15, we encountered a race condition where the
replication LSN was closed, but we still tried to use it to send
definite errors from the primary source (e.g. the publication was
dropped) to the subsources.

Instead, we can invent a new LSN to send along the definite errors.