sql: Fix panic on correlated casts
Previously if a cast from one type to another type was implemented
using SQL, and the SQL contained a SELECT or other correlated
expression, then the cast would cause a panic in certain situations.
The panic would happen if the cast was involved in any of the following
casts:

- String to Array
- String to List
- String to Map
- String to Range
- Record to Record
- Array to Array
- List to List

These casts incorrectly assumed that the cast implementation did not
contain a correlated expression.

This commit removes the panic from these casts, but does not actually
implement the casts.

Fixes #19052