testdrive: Fortify subscribe-output.td
The test used FETCH ALL, which is not deterministic as ALL means
the data that is available as of the FETCH, which may or may not
include a progress marker.

Switch to FETCH 2 to make sure the test will block until the
expected progress marker is emitted by the cursor.