Handle lazy funcs by checking errors explicitly
Previously, we assumed that if an expression was a function call, and
one if its arguments failed, then the overall expression would fail.
While this is mostly true, this is not true for `AND` and `OR`. Instead
of hardcoding these two functions (brittle!) let's teach `flat_map` to
actually test the behaviour.

We additionally add a test for non-strict `OR` as a regression test.