docs/developer: make design doc lifecycle more "targeted"
This should bring it in line with current practice: at our scale (and
when growing more), it's not feasible to "just throw out" a design
doc and assume the relevant people will find it.

This makes it more explicit that you should identify the stakeholders
and get explicit approval (some form of it) from them, instead of an
approach based on timeouts.