compute: always use the simple JoinThinker strategy
This commit adjusts the `JoinThinker` to always use the simple
nested-loops strategy to produce join matches, regardless of the number
of edits. The rationale is that the more complex strategy is only
beneficial when there are many edits at different times, which is
generally not the case in Materialize where times are usually at second
granularity. Using the simple strategy makes it easy to yield in the
middle of matching, which is not the case with the complex strategy.

Getting rid of the complex matching strategy also allows us to delete
the `ValueHistory` and `HistoryReplay` types.