compute: remove EditList abstraction
This commit removes the `EditList` type and instead makes the join
matching work directly on the batch/trace cursors. `EditList` was useful
to support replaying the edit history, but we don't do that anymore.
Doing the matching directly on the cursors allows us to delete some
code, but more importantly makes it obvious at which points
`Deferred::work` is allowed to yield.