adapter: Fix AS OF panics in transactions
Previously, executing `AS OF` in a read only transaction could trigger
a panic. This was because the timestamp selected for the `AS OF` would
conflict with timestamps selected in previous queries.

This commit fixes this panic by not comparing the timestamp used in
`AS OF` queries against the timestamps of other queries in the same
transaction.

Fixes #19180
Fixes #19178