sql: Improve ergonomics of setting current schema
Previously, we followed the PostgreSQL syntax for how to set the
current search_path. That is a user could type
`SET search_path = <list-of-schemas>` or `SET schema <single-schema>`.
`TO` also works as a replacement for `=` in the first example. Note the
lack of an `=` in the second example. The reason PostgreSQL does this
is that when using parameter aliases (such as `SCHEMA`) the user has to
omit the `=`/`TO`. This differs from how we treat databases and
schemas. For those objects users can type
`SET database = <database-name>` or `SET cluster = <cluster-name>`
respectively. Many users were confused of the differences between
schemas and databases/clusters.

This commit improves the ergonomics of setting the current schema by
allowing the syntax `SET schema {= | TO} {ident | string}` which has
the same result as `SET schema {string}`. It also adds `SHOW schema`
which acts as an alias for `SELECT current_schema()`.

Additionally, `SHOW database`, `SHOW cluster`, and `SHOW schema` emit a
notice if they are unable to display a resolvable object.

Resolves #18946