MIR type checking (#18666)
An MIR typechecker that goes beyond `col_with_input_cols` to do thorough internal consistency checking. As we consider changes to the MIR optimizer, having thorough consistency checking will help us catch corner-case regressions before they manifest as runtime failures.

The type checker is run at the beginning and end of each phase of the optimizer; as the checker runs, it collects type information for the plans corresponding to every global ID. When the type checker re-encounters a global ID's optimized plan, it checks the optimized plan for consistency and confirms that the new plan's type agrees with the
old one.

Note that this type comparison is about the external, visible type of the expression---internally types can change radically, but the visible type of a global ID is treated as stable.

The typechecker is imprecise in a few ways:

- Not all globals are accounted for.
- Some nullability information on local (let-bound) variables is incorrect.
- Join equivalences should always have at least one non-nullable field, but that invariant isn't upheld by the optimizer and/or the typechecker is not precise enough to notice.

If you get any type errors, noisy logs, or have other questions about the typechecker, please use https://github.com/MaterializeInc/materialize/issues/19101 to discuss them.