testdrive: Fortify ubscribe-output.td (take #2)
FETCH N returns "at most N" rows, which causes test flakiness
when used in conjunction with WITH PROGRESS, as the timestamp
row may or may not be available at the time the FETCH runs
and FETCH N will not wait for exactly N rows to become available.

Attempt to fix the test by not waiting for the second timestamp
row to be emitted at all. The test already checks with the
previous query that at least one timestamp row has been
emitted.

Fixes #19267