In-place consolidation in consolidate_named_if
This commit introduces in-place consolidation with a ConsolidateBuffer in timely-util's consolidate_named_if. The function is used for conditional consolidation in rendering, especially in monotonic one-shot SELECTs. It was observed in MaterializeInc/materialize#18890 that direct consolidation leads to potentially limited parallelism when input domains are limited, which this change aims to ameliorate by performing per-worker, pipelined consolidation.