Arranged per-worker preaggregation in consolidate_named_if
This commit changes the consolidate_named_if function provided in timely-util to perform the consolidation in two stages. Firstly, a pipelined arrange step achieves per-worker preaggregation of differences. Secondly, a regular consolidate_named enforces data exchange and executes the final aggregation. The scheme is intended at improving the paralellism profile of forced consolidation in monotonic one-shot SELECTs, albeit at a higher overhead and memory cost. It is expected that the overheads of the newly introduced first step be manageable where the step is not helpful, but pay off in queries where parallelism can be recouped. The memory cost is non-trivial and equals about twice of the needs for a direct consolidate_named. However, these extra memory requirements only manifest in monotonic one-shot SELECTs, and are thus expected to be a temporary spike when forced consolidation is necessary.