Incorporated feedback regarding refinement pass
This commit revisits the proposal in the plan refinement pass to use an analysis approach based on the Interpreter trait rather than an iterative approach (with either poor complexity or analysis imprecision). The proposal is now bolder, but there seems to be agreement in that it is a better path forward.