Use twisted tabulation hashing in consolidate_named_if
This commit moves away from a radix-based approach in consolidate_named_if and adopts instead a re-hashing of the output of DD's hash function on Exchange. The re-hashing is performed by relying on a crate that implements twisted tabulation hashing. The latter function is practical, but provides good bounds for balls-into-bins problems when the number of bins is small (as is our case). A proof is given by Mihai Patrascu, Mikkel Thorup: Twisted Tabulation Hashing. SODA 2013: 209-228, available at https://epubs.siam.org/doi/epdf/10.1137/1.9781611973105.16.