frontegg-client: enables authentication using email and password (#19421)
The motivation for this PR is to enable authentication using email and
password. It is a feature needed by `mz` to login a user without using
the web browser.

Currently the `frontegg-client` supports to authenticate using
AppPasswords, so an enum has been added: `Authentication` that could be an
`AppPassword` or `Credentials`, and support for each type of request.