Use AHash in consolidate_named_if
This commit switches from the use of DD's default FnvHasher followed by twisted tabulation hashing to the use of AHash directly instead. The latter hash function is empirically demonstrated to strike a good speed vs. quality trade-off and is already heavily used by our dependencies. Therefore, it seems to be a better choice to avoid balance problems in exchanges, particularly the one for conditional consolidation in one-shot SELECTs.