adapter: rename Command::{Describe => Prepare}
@umanwizard pointed out that `Command::Describe` corresponds to `Parse`
in the pgwire protocol, not `Describe`. Rename the coordinator command
to avoid confusion. The new name, "prepare", does not directly map to
any pgwire commands, and seems to clearly map to what the command does:
turning a parsed SQL statement into a named prepared statement.

Note that "describe" remains in use in the coordinator/SQL planner for
the functions that convert parsed SQL statements into "statement
descriptions". This seems appropriate, as those descriptions are what
the `Describe` pgwire command later references.