Early exit when variadic evaluation is expensive
In general, to interpret a function, we evaluate it (using the normal
evaluation code) with carefully-chosen arguments to infer its behaviour.
This is normally cheap - for a binary function at worst about 16
evaluations - but can get expensive for a variadic function where it is
worst-case exponential in the number of arguments.

We can optimize common cases to avoid hitting the exponential path, but
it's ~impossible to avoid in general, so let's add this conservative
fallback.