adapter: Try and prevent duplicate terminates
Very rarely we have seen errors where the Coordinator tries to
terminate a session, but the connection ID is unknown to the
Coordinator. This ends up causing a panic.

One potential explanation is that a session tries to terminate itself
more than once. This commit attempts to prevent duplicate terminates in
a very specific scenario. Specifically the duplicate terminate may
happen during a panic unwind, if the panic happens right after sending
a terminate command but before cleaning up the session.

Related to #18996