Handle float over/underflows in pow, exp
The SQL spec says about pow:

    [...] the result is the result of
    EXP(NVEE*LN(NVEB))

and about exp:

    If <exponential function> is specified, then let V be the value of
    the simply contained <numeric value expression>.  [...] the result
    is e (the base of natural logarithms) raised to the power V. If
    the result is not representable in the declared type of the
    result, then an exception condition is raised: data exception —
    numeric value out of range.

Arguably, the result is representable, since inf is part of the type,
but the section on "characteristics of numbers" seems to indicate that
loss of the most-significant digit is considered an overflow.

Notably, this change matches Postgres's behavior for these functions.

In this case (unlike some of the other functions), Rust's f64
implementations compile to the LLVM intrinsics rather than libm calls,
and the former don't check any of these error cases for us.  (Then
again, historically libm implementations have been buggy enough that
lots of code duplicates those checks anyway...)

Fixes #19275.