persist: fix pushdown metrics
These were on ShardMetrics an attempt to make them per-shard, but for
SELECTs, there's no good way to make that ShardMetrics live long enough
for prometheus to scrape it. This is a footgun (as evidence by me
spending a bunch of time debugging it XD), so fix it by making them not
per-shard for now.

While I'm in here, also add some metrics for audit, so we can keep an
eye on how much work it's causing the system to do.