Allow setting DD's idle_merge_effort to None
This commit adapts the treatment of the CREATE CLUSTER REPLICA option IDLE ARRANGEMENT MERGE EFFORT to allow for a NULL value to be provided. When NULL is given, a None value will be provided to DD's idle_merge_effort parameter. This change is significant because the semantics in DD of an idle_merge_effort equal to Some(0) differs from an idle_merge_effort equal to None. The former implies some effort is still applied, while the latter effectively turns idle arrangement compaction off. At a high level, one can conceptualize DD's option as advisory: Setting an effort value implies that there will be an attempt to apply, if possible, some effort greater than or equal to that value per exertion cycle.