[persist/expr] Fallback and optimize for worst-case interpreter path (#19448)
The `ColumnSpecs` abstract interpreter has a exponential-time worst-case
for function evaluation. In brief: each argument has perhaps four
"special" values -- null, errors, range bounds -- and the interpreter
needs to evaluate the function with every combination of those values
for each argument, which makes the overall runtime exponential in the
length of the argument list. Normally this is fine: for binary
functions, for example, the exponent is 2, and this is still plenty
fast. But this can get expensive for variadic functions with
medium-to-long argument lists.

This PR addresses the issue in two ways:
- Adding a new linear-time implementation for associative functions. (If
a function is associative, we effectively re-associate it to be
evaluated as a linear number of binary function calls, each of which is
sufficiently fast.) Conveniently, functions already have an
`is_associative` method, and all the functions we care about are
included.
- Limiting the number of arguments for the non-optimized case. This
prevents both deep stacks (as seen in
https://github.com/MaterializeInc/materialize/issues/19399) and lots of
calls to the function itself.

So the common case is now fast, and the uncommon case makes the
approximation worse but does not blow up.

Co-authored-by: Philip Stoev <pstoev@materialize.com>