timely-util: synchronize operator shutdown between workers
This PR fixes a longstanding correctness issue with token based dataflow
shutdown (see #18837 for details). Before this change, the tokens
returned by the AsyncOperatorBuilder were purely local tokens which when
dropped had immediate effect to the affected operators in the worker.
Specifically, when the old tokens were dropped the operator logic was
immediately dropped, causing the frontiers to advance.

This works fine in a single-worker setting but fails in a multi-worker
settings. This is because while the worker that dropped the token knows
not perform any unwanted sideffects (e.g incorrectly advancing a persist
shard's frontier) the rest of the workers do not, and if the other
workers receive the frontier notification before they have also been
notified to shutdown, bad things happen.

The solution implemented in this PR changes the tokens to no longer be
worker-local. Instead, each operator token constructs a timely channel
in order to communicate with the same operator token in the rest of the
workers. By doing so each token can know both if the local worker has
dropped its token and whether *all* workers have done so. This is
accomplished by simply counting the number of button presses and
concluding that all workers have pressed the button once the count
reaches the number of peers.

Each tokened operator will immediately stop polling its logic future (or
stop stop draining its input handles if it's in that state) when its
token is locally dropped. After that point it will wait until all
workers have also stopped scheduling their logic future before going
ahead with dropping the logic future which might start producing bogus
frontier updates.

The end result is that "weird" results are only produced when all
workers have made the positive statement that they have paused their
computation.

Fixes #18837