persist: remove soft-assertions-based audit
When we hooked this up, the `system_parameter_defaults` stuff wasn't
totally wired up in CI yet. We wanted to be very sure that we were
getting auditing in all CI jobs. Now that QA is trying to write tests to
verify the performance benefit of pushdown, pull out the
soft-assertions-based audit trigger and rely entirely on
system_parameter_defaults to set auditing in CI.

Closes #19571