limits test: Expand the size of the Mz instance tested
With PubSub enabled, CRDB is no longer the limiting factor, so
it is possible to expand the instance-size workflow in various
dimensions.

At the same time, we are reducing the number of replicas to 3
per cluster. This will match reality closer in the sense that
users are expected to have two replicas for failover as well
as posssibly an occasional third replica during maintenance
operations such as scale-up or scale-down of the compute instance
type.

Relates to #19246