Structure for single-time consolidation refinement
This commit introduces implementations for Interpreter and Transform
that are then used in an additional LIR-level plan refinement step. The
ultimate goal of this refinement step is to relax the must_consolidate
flag whenever possible in single-time dataflows. The relaxation depends
on analysis of physical monotonicity, in a single-time context, of the
inputs given to the monotonic Plan nodes that are relaxation
candidates. The introduced implementations at this point are inert: No
transformation is attempted, and analysis returns always the most
conservative outcome. The commit simply ensures that the code structure
is properly adapted, and provides basic implementations for necessary
data representations such as PhysicallyMonotonic.