Implement SingleTimeMonotonic Interpreter
This commit provides an implementation for an Interpreter that derives
single-time monotonicity judgments of LIR plans. In addition, the
commit makes adjustments to the Context tracked in the LIR abstract
interpretation framework's Fold and FoldMut so that recursive contexts
can be reasoned about. Finally, adjustments to the Transform and
BottomUpTransform APIs are carried out to adjust TransformConfig. The
latter allows for the base set of physically monotonic IDs to be
supplied to the analysis.