storage: don't resolve remote hosts when using ssh tunnels
The `openssh` crate has a bug in which it attempts to resolve the target
host, which my only be resolvable on the SSH bastion host, on the
client.

Upgrade to a patched version of `openssh` that does not have this bug,
and add a test for this scenario (that would fail without the patched
version of `openssh`).