ore: remove Deref implementation from IdAlloc
I added this Deref implementation in 6a9524b, but it turns out
@ParkMyCar intentionally omitted the Deref implementation to make it
more obvious when someone is fetching the raw value inside of the
`IdHandle`, as it's a somewhat dangerous operation.

So, this commit is a spirtual revert of 6a9524b, except that it goes a
bit further than the original code to prevent accidental misuse of a raw
ID.  The method to retrieve a raw ID from an ID handle is now called
"unhandled" instead of "val", to make it more obvious that the return
value must be handled with care.