adapter: add `Command::PrivilegedCancelRequest`
When Materialize itself is generating a cancellation request (rather
than an external client), there is is no need to require the secret key
for the connection. Instead, we can demand a `ConnectionId` handle.

This makes it easier to reason about the correctness of
SessionClient::send_with_cancel; it is now obvious that the connection
ID remains valid while the coordinator processes the cancellation
request.