Separate out preparation stage in sqllogictest run_query
This commit refactors the method run_query of RunnerInner in sqllogictest to break it down into a prepare_query stage and an execute_query stage. Query preparation performs parsing and transaction setup, while query execution issues the query and validates the results. The refactor does not change the current functionality, but rather just arranges the code to facilitate extension.