auto_index_selects argument in sqllogictest
This commit introduces the argument auto_index_selects in sqllogictest. The argument is passed along in the implementation to RunnerInner, its point of use. However, the logic to use this argument is not provided in this commit, so trying to employ the argument simply results in a panic. This will be rectified in subsequent commits.