storage: compact reclock trace on handle drop
This PR is fixing a serious bug of the reclocking implementation that
when triggered can produce incorrect data.

The `ReclockFollower` object can have multiple handles to it. Each
handle carries a handle-specific since frontier that can be used for
downgrades. The lower bound of all those since needs is what drives the
compaction of the in-memory trace of remap bindings.

While the logic described above was correcty managed in the
`ReclockFollower::compact` method, that was not the case when handle was
dropped. In that case only the frontier was being updated without also
compacting the trace. This is an invariant that the rest of the methods
rely on and when broken can lead to reclocking data to earlier
timestamps than it should.

The fix is to just call `self.compact({})` on `Drop` instead of doing
anything special there. A test is included that demonstrates how the
issue manifests.