sql: Refactor ObjectType to be more accessible
The ObjectType enum is used throughout the code base in multiple
crates. However, it's defined in the mz_sql_parser crate which is
purposely kept isolated from many crates. This can make working with
the enum a bit cumbersome.

This commit creates a second ObjectType enum in the mz_sql crate. All
usages of the ObjectType enum outside the parser use the new mz_sql
enum. The mz_sql_parser enum is always converted to the mz_sql enum
in the planner, which serves as a boundary between the mz_sql_parser
crate and the other crates.

The specific motivation for this refactor is that we'd like to store
ObjectTypes in the stash, without adding mz_stash as a dependency for
mz_sql_parser or adding mz_stash as a dependency for mz_sql_parser.