stash: Add default privileges stash collection (#19796)
This commit adds default privileges to the stash. The stash collection
is pre-populated with a single default privilege that grants USAGE
privileges on all types created to the PUBLIC role. Currently, the stash
collection is not used, a later commit will read and update the values
in the collections.

Works towards resolving #19647