Store JSON numeric stats as ProtoDatums
Previously, JSON numeric stats were stored as floats. It's fine that
stats were a little lossy, because we only need rough bounds on the true
value. However, it was difficult to guarantee that the bounds were
correct, and that roundtripping the numeric bounds through float made
them strictly wider.

Also, for top-level numerics we just use the ProtoDatum encoding, which
is lossless and ~fine. So let's use that here too.