adapter: Add infrastructure for default privileges (#19793)
This commit adds default privileges to the stash, system catalog tables,
and the in memory catalog. When a new object is created, the default
privileges are automatically applied to the object.

Currently, there's no way for a user to modify the default privileges,
it's only pre-populated with some system default privileges. In a later
commit we will add the `ALTER DEFAULT PRIVILEGE` command to allow users
to alter these privileges.

Works towards resolving #19647