Avoid divergence of hash state in consolidation
This commit fixes a correctness bug in the usage of the hashing API, in particular AHash's, in the Exchange pact of mz_consolidate_if. Additionally, the commit extends the patched usage of the hash function to mz_consolidate, thus increasing our confidence on good distribution of rows to workers given the use of a higher quality hash function than the DD's default FnvHasher. The bug was caused by: (a) initializing the hash function with a method that would employ randomness, thus leading to divergent hash state across workers; (b) not reinitializing, but reusing the hash function across invocations, thus computing from a hash state that was not well-defined.