Set default replica config with DEFAULT_IDLE_ARRANGEMENT_MERGE_EFFORT
This commit ensures that newly created replicas that obtain a default replica configuration will pick up the default value of the idle_arrangement_merge_effort setting. Note that previously the default configuration just left this setting as None and the code would reinterpret it as Some(DEFAULT_IDLE_ARRANGEMENT_MERGE_EFFORT). Since the latter is no longer done, it is necessary, in addition to this commit, to write a one-time migration of previous None values in the stash to Some(DEFAULT_IDLE_ARRANGEMENT_MERGE_EFFORT). The migration will be provided in a later commit.