adapter: Avoid duplicate role membership revokes
When either the group role or the member role of a role membership is
dropped, then the role membership is automatically revoked. For example
if role r1 is a member of role r2, if either r1 or r2 are dropped, then
we automatically revoke r2 from r1.

Previously, when both r1 and r2 were dropped in the same statement, we
would try to revoke r2 from r1 twice, causing a panic. This commit
fixes that issue so the role membership is only ever revoked once.