persist: remove in-mem blob cache impl
We believe that the moka library used to implement it is causing
segfaults on the release canaries. It'd be nice to figure out how to
reintroduce this at some point, so leave the metrics and LD configs
around.

We could instead leave this impl and disconnect it, but there's some
appetite for removing the dep and it's easy to pull out of git history
if the current code ends up being useful.