Add process orchestrator wrapper support to sqllogictest
This change adds a --orchestrator-process-wrapper parameter to sqllogictest
inspired by the parameter of the same name to `environmentd`. This allows
for wrapping calls to `clusterd` with anther command, such as a profiler.

Also send a SIGTERM to child processes before killing to give the wrapper
a moment to cleanup its state, or write down pending data to disk. This
makes it possible to use `perf` as a wrapper.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>