environmentd: remove waiting_on_leader_promotion testing knob
This commit removes the `waiting_on_leader_promotion` testing knob from
environmentd configuration. The knob was added to support the new
upgrade orchestration in #19754 so that a Rust unit test could ascertain
the server's internal HTTP address before the server was fully booted.

We prefer to avoid testing knobs whenever possible, and this one turns
out to be easy to avoid with a small refactoring to the environmentd
server initialization API. The new API is split into two functions: one
for creating the listeners, and one for booting the server from those
listeners. The caller can ask the listeners for its bound addresses
before booting the server.