sql: introduce CastContext::Coerced
The PG regclass types have different behavior if they're
explicitly cast from strings, vs. if they're coerced from string
literals. This suggests a fourth type of cast context, so
introduce it.

Because checking that the caller provided the appropriate
CastContext for the cast, derive Ord, PartialOrd, as well to
simplify the code.