stash: Add system privileges stash collection
This commit adds system privileges to the stash. The stash collection
is pre-populated with a single system privilege that grants all system
privileges to the mz_system role.

Currently, the stash collection is not used, a later commit will read
and update the values in the collections.

Works towards resolving #19997