adapter: Add infrastructure for system privileges
This commit adds default privileges to the system catalog tables and
the in memory catalog.

Currently, there's no way for a user to modify the system privileges,
it's only pre-populated with some default system privileges. In a later
commit we will add the ability to alter system privileges via the
`GRANT` command.

Works towards resolving #19997