Fast SQL Logic Tests with auto_index_selects
This commit adds a selection of SLTs to the fast SQL Logic Tests in CI that are run with the option --auto-index-selects. The tests take longer to run, so the selection is kept small. Since under auto_index_selects the SELECT statements are executed redundantly as one-shot and through an indexed view, the tests selected can be excluded from the normal run of sqllogictest.