Bundle execution finalization state into `ExecutionContext`. (#19881)
We pass a `tx` and `session` around various functions as part of the
statement logging lifecycle; at the end of statement execution, we need
to send the `session` back through the `tx` so that the `pgwire` layer
can have access to it again.

As part of statement logging, there will be some more tasks we need to
perform at the end of each execution, so as a precursor to that, I
bundled that logic into one struct, `ExecuteContext`, which must be
"retired" at the end of statement lifecycle. This involved replacing
every `tx.send((session, result))` with a call to either `retire` or
`retire_async`.

Each of these functions returns the `session` back through the `tx`, and
then notifies the coordinator so that it can perform any other
cleanup/logging work as necessary. The difference between them is that
`retire` requires `&mut` access to the coordinator, whereas
`retire_async` communicates with it by sending a message.

Currently, the extra work that the coordinator does is nothing, but that
will change soon.

For more details, see the comments on `ExecuteContext` and
`ExecuteContextExtra`.