storage: skip calling finalize_shards
We have an active incident where it appears that shard finalization has
picked up some very old shards (even as far back as October) and is
attempting to gc as part of finalization. Work around this for now by
disabling shard finalization.

Hopefully the incremental gc PR that just went in to main will let us
work through this backlog once it's out.