compute-client: peek metrics
This commit adds metrics for finished peeks:
 * the total counts of peeks performed since the last restart
 * a histogram of peek durations

Both metrics are labeled by peek result type (completed or canceled),
since the difference will likely be interesting to us. For example, if
we want to calculate the average peek response time, we'll probably want
to exclude canceled peeks as they might artificially lower the average.