compute: introduce a `CollectionState`
Prior to this commit, the per-collection state tracked by compute was
spread over several `BTreeMap`s in `ComputeState`:
 * `sink_tokens`
 * `sink_write_frontiers`
 * `flow_control_probes`
 * `reported_frontiers`

This makes it hard to keep track of the collection state as a whole and
invites bugs caused by inconsistencies (e.g. we might forget to drop all
of the state for a collection).

This commit refactors the `ComputeState` by replacing the four
per-collection fields mentioned above with a single `collections` map.
Values in this map are `CollectionState` objects that each contain the
entire state specific to a single collection. An exception is trace
state, which is kept in the `TraceManager` and pulling that apart seemed
too much of a hassle.