doc/user: document `FORMAT JSON` for sources
`FORMAT JSON` is a new source format that decodes Kafka messages into a
single column named `data` with type `jsonb`.

We have more to do to support automatically enforcing schemas on the
JSON (see issue #7186), but the current `FORMAT JSON` feature is useful
enough on its own to be worth enabling and documenting.

Fix #18979.