doc: Copy anchor links to clipboard on click
We use Anchor.js to add anchor links to headings so we
can link users to specific sections of a page.

This PR adds a bit more JavaScript that copies anchor
link URLs to the clipboard when the anchor icon is
clicked. This makes it easier to share anchor links.

Adapted from https://dev.to/tqbit/how-to-use-javascript-to-copy-text-to-the-clipboard-2hi2.

Fixes: #16876