persist: introduce an id for each SpineBatch as it walks through levels
- Each batch is given an incrementing integer id as it is pushed/inserted
  into the persist Spine.
- This is represented in a half-open interval and called `SpineId`:
  `[0,1), [1,2), ...`.
- We then maintain an invariant that the SpineIds for batches in the
  spine all "line up" (in the same way that the lowers and uppers do)
  by, whenever batches are merged/compacted, giving the new batch the
  SpineId interval spanning the inputs. E.g. `[4,7)+[7,9)=>[4,9)`.

This allows us to uniquely identify a batch as it walks up (and down!)
the spine levels.

The immediate motivation is a follow-up PR to roundtrip the internals of
spine directly instead of the set of HollowBatches. This latter has
caused all sorts of complexity (see the various `apply_diffs_spine`
hacks) and bugs (#17214, among others) because two copied of persist
State might end up with a different set of HollowBatches depending on
which SeqNo they start replaying persist state from. I now consider
round-tripping the HollowBatches to be a mistake in hindsight.

This id scheme also helps set up a number of more ambiguous compaction
2.0 goals, like generalizing Spine to have more than one batch in a
level and/or compacting more than two inputs at at time.