adapter: Fix mz_introspection cluster migration
A previous commit added a migration to grant the mz_introspection role
USAGE privileges on all existing clusters. It set the grantor of this
privilege to mz_system. This was incorrect, the grantor should have
been, the owner of the cluster. This commit fixes the migration to be
the owner of the cluster.