Enable monotonic oneshot SELECTs in fast SLTs
This commit turns on the feature flag enable_monotonic_oneshot_selects in the present set of SLTs that are run with --auto-index-selects. The set may change in the future, so we can also think of a complementary strategy wherein the feature is activated whenever --auto-index-selects is provided. This commit, however, has value in that a manually validated set of tests will now be subsequently be run with the feature turned on in CI.