SKETCH: we should no longer need the idempotency info
This code made sure that requests were idempotent for at least 15
minutes - the length of a writer lease - but we now maintain idempotency
by checking the states in consensus which are now also kept around for
at least 15 minutes.

If we want to do this for real, we can remove a lot more code. But this
is hopefully enough to illustrate that we're not relying on leases in
the CAS path.