sql: Prevent SHOW commands in views
When a view is created with a SHOW command, the create SQL definition
will have the explicit `SHOW` command in it. Many `SHOW` commands will
expand into a sub-query that involves the current schema of the
executing user. When Materialize restarts and tries to re-plan these
queries, it will only have access to the raw `SHOW` command and have no
idea what schema to use. As a result Materialize will fail to boot.

Some `SHOW` commands are ok, like `SHOW CLUSTERS`, and there are
probably other ways around this issue. Such as expanding the `SHOW`
command in the SQL definition. Or explicitly using an unmaterializable
function in the HIR to prevent problematic views. For now, we just
disallow any `SHOW` commands in views since the issue exists in prod
and can corrupt catalogs.

This is a potentially breaking change since some existing views with
`SHOW` might exist in production.

Fixes #20188