Include ahash with default_features=false
This commit changes the dependency on AHash in timely-util's Cargo.toml to be explicit about avoiding default features. This specification was lost as part of the backport in #20124. It is not strictly necessary for the fix related to AHash usage in timely-util, but it is a good practice so as to steer the build of AHash in the right direction (i.e., avoiding runtime random number generation).

The commit also includes some minor improvements to code comments related to AHash's selection of either its primary hasher, which depends on AES instructions, or the fallback hasher. It was observed in the review of #19848 that the target features we get enabled by default do not include AES. Further evaluation of the possibility to do so is the subject of #20028.