storage: ldbc load generator
This load generator downwloads the ldbc-snb-bi dataset from github. It
currently doesn't retry or expose any errors. Users will need to drop
and re-create if they believe an error has occurred. (This may be fixed
in the future, but is good enough for now.)

The added testdrive file is not currently tested anywhere, but passes
locally.