ci/test: build ARM only on main
The ARM build is stubbornly slower than the x86_64 build. Since we're
not actively running Materialize on ARM in production right now, stop
running the ARM build on PRs, and switch to running it only on main.
This way we'll still quickly catch bugs that cause us to stop compiling
on ARM, but we'll spend less money and waste less developer time in
doing so.